move file and created new folders
