+++ pipeline/preprocessing/df_processing.py
@@ -1,17 +1,10 @@
-""" Script to intially explore our json data.
-    for now, we only use the create_df_from_json and 
-    calculate_distance_and_speed function in the pipeline.
+""" Used to calculate speeds and distances for a dataframe.
 """
 
-from functools import reduce
 import json
 from math import sqrt
-from itertools import tee as copy_iterable
-from statistics import median, mean
-import sys
-from typing import Iterator
-import pandas as pd  # type: ignore
-
+import pandas as pd
+from archive.analysis_metrics import calculate_metrics, filter_segments  # type: ignore
 
 def shift_elems(l: list) -> list:
     """shift elemets of list, ie:
@@ -28,8 +21,6 @@ def shift_elems(l: list) -> list:
 
 
 coordinate = tuple[float, float]  # type alias
-
-
 def calc_utm_dist(utm1: coordinate, utm2: coordinate) -> float:
     """calculate the distance between two utm coordinates
     uses formula for euclidian distance, which is not 100% accurate?
@@ -91,63 +82,6 @@ def ms_to_kmh(speeds: list[float]) -> list[float]:
         list[float]: new list of speeds
     """
     return [speed * 3.6 for speed in speeds]
-
-def verify_solution(l: tuple[Iterator[bool], Iterator[bool]]):
-    """ Auxiliary fuction used in filter_segments, to verify the solution works.
-    I assume that trips do not loop back and go through the same
-    segment more than once. This function is a sanitity check for this.
-
-    Args:
-        l (list[bool]): list with boolean mask of all coordinates that are in
-        the segment of interest.
-    """
-    safe_cmp = lambda b, acc: True if len(acc) == 0 else acc[-1] != b
-    reduced: Iterator[
-        bool
-    ] = reduce(  # convert [False, False, True, True, True, False] -> [False, True, False]
-        lambda acc, bool: acc + [bool] if safe_cmp(bool, acc) else acc, l, []  # type: ignore
-    )
-
-    # if there are more than one true in list
-    # then the car has looped and break assumption
-    if len(list(filter(lambda e: e is True, reduced))) > 1:  # type: ignore
-        print("Bad assumption...")
-        sys.exit()
-
-def filter_segments(df: pd.DataFrame, osm_id: int) -> pd.DataFrame:
-    """Discard all rows from coordinates that do not correspond to the given osm_id.
-
-    Args:
-        df (pd.DataFrame): geodata dataframe
-        osm_id (int): segment id
-
-    Returns:
-        pd.DataFrame: dataframe only with coordinates corresponding to given osm_id
-    """
-    def filter_func(row):
-        """read above doc string
-        """
-        valid_osm_ids = map(  # id mask corresponding to which coordinates to keep
-            lambda elem: elem == osm_id, row["osm_id"]
-        )
-        verify_solution(copy_iterable(valid_osm_ids))
-
-        valid_coordinates = list(
-            map(
-                lambda tup: tup[0],  # only keep seconds list containing the coordinates
-                filter(  # only keep valid coordinates
-                    lambda elem_and_is_valid: elem_and_is_valid[1],
-                    zip(
-                        row["coordinates"], valid_osm_ids
-                    ),  # combine coordinates and valid som_ids
-                ),
-            )
-        )
-        row["coordinates"] = valid_coordinates
-        return row
-
-    return df.apply(filter_func, axis=1)  # type: ignore
-
 
 def calculate_distance_and_speed(df: pd.DataFrame):
     """Calculate the distance between each coordinate.
@@ -211,19 +145,6 @@ def calculate_speeds(df: pd.DataFrame) -> None:
     df.drop(["time_difference"], axis=1, inplace=True)
 
 
-def verify_outlier(filtered_df: pd.DataFrame):
-    """Used to verify the outlier from segment_10240935_linestring.json.
-    The speed calculated was initially wrong. But now fixed. 
-
-    Args:
-        filtered_df (pd.DataFrame): dataframe
-    """
-    print(filtered_df.iloc[7].time_difference)
-
-    for (x, y, time), speed in zip(
-        filtered_df.iloc[7]["coordinates"], filtered_df.iloc[7]["speeds"]
-    ):
-        print(time, speed, sep=", ")
 
 def main():
     filename = "segment_10240935_linestring.json"
